fix albumart race condition
Ensure the image profile is not applied to cover art. I believe that
what happens very rarely is a race condition in auto_profiles.lua where
it receives one of the 2 update notifications and re-evalutes the
profile before receving the other, so it is fixed by observing only
current-tracks/video.

In d85ad47 I thought it depended on replacing current-tracks/albumart
with current-tracks/audio, but it kept happening after it.